requirements.txt
aniso8601==3.0.2
click==6.7
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dotenv==0.8.2
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.9
Werkzeug==0.14.1
